Revert "Merge branch 'loudnate/master' into dev"
This reverts commit 8ce57ba2718e185ca53422588d6aafa03d5b8499, reversing
changes made to e546c9b8fd453d28775a06689f0a6d06de3d12b8.

+++ Loop/Extensions/PumpModel.swift
@@ -14,7 +14,7 @@ extension PumpModel {
         return Int(rawValue)! / 100
     }
     
-    var reservoirCapacity: Int {
+    var reservoirCapacity: Double {
         switch size {
         case 5:
             return 176
